Prohibit deletion of words from default databases

+++ Word Bomb/UI/Database/DatabaseView.swift
@@ -62,7 +62,11 @@ struct DatabaseView: View {
                     
                     Button(action: {
                         withAnimation {
-                            db.remove(wordsToDelete)
+                            if db.isDefault_ {
+                                Game.errorHandler.showBanner(title: "Deletion Prohibited", message: "Cannot modify a default database!")
+                            } else {
+                                db.remove(wordsToDelete)
+                            }
                         }
                     }) {
                         Image(systemName: "trash")
